// src/modules/checkout/usecase/place-order/place-order.usecase.ts
import UseCaseInterface from "../../../@shared/usecase/use-case.interface";
import {PlaceOrderInputDto, PlaceOrderOutputDto} from "./place-order.dto";
import {ClientAdmFacadeInterface} from "../../../client-adm/facade/client-adm.facade.interface";
import ProductAdmFacadeInterface from "../../../product-adm/facade/product-adm.facade.interface";
import Product from "../../domain/product.entity";
import StoreCatalogFacadeInterface from "../../../store-catalog/facade/store-catalog.facade.interface";
import Id from "../../../@shared/domain/value-object/id.value-object";
import Client from "../../domain/client.entity";
import Order from "../../domain/order.entity";
import {PaymentFacadeInterface} from "../../../payment/facade/payment.facade.interface";
import CheckoutGateway from "../../gateway/checkout.gateway";
import {InvoiceFacadeInterface} from "../../../invoice/facade/invoice.facade.interface";

export default class PlaceOrderUseCase implements UseCaseInterface<PlaceOrderInputDto, PlaceOrderOutputDto> {

    constructor(
        private readonly _clientFacade: ClientAdmFacadeInterface,
        private readonly _productFacade: ProductAdmFacadeInterface,
        private readonly _catalogFacade: StoreCatalogFacadeInterface,
        private readonly _repository: CheckoutGateway,
        private readonly _invoiceFacade: InvoiceFacadeInterface,
        private readonly _paymentFacade: PaymentFacadeInterface,
    ) {}

    async execute(input: PlaceOrderInputDto): Promise<PlaceOrderOutputDto> {
        const client = await this._clientFacade.find({id: input.clientId});

        if (!client) {
            throw new Error("Client not found");
        }

        await this._validateProducts(input);

        const products = await Promise.all(
            input.products.map((p) => this._getProduct(p.productId))
        );

        const myClient = new Client({
            id: new Id(client.id),
            name: client.name,
            email: client.email,
            address: client.address.street,
        });

        const order = new Order({
            client: myClient,
            products,
        });

        const payment = await this._paymentFacade.process({
            orderId: order.id.id,
            amount: order.total,
        });

        const invoice = payment.status === "approved"
            ? await this._invoiceFacade.generateInvoice({
                name: client.name,
                document: client.document,
                street: client.address.street,
                complement: client.address.complement,
                number: client.address.number,
                city: client.address.city,
                state: client.address.state,
                zipCode: client.address.zipCode,
                items: products.map((p) => ({
                    id: p.id.id,
                    name: p.name,
                    price: p.salePrice
                })),
            })
            : null;

        payment.status === "approved" && order.approved();
        await this._repository.addOrder(order);

        return {
            id: order.id.id,
            invoiceId: payment.status === "approved" ? invoice.id : null,
            products: order.products.map((p) => ({ productId: p.id.id })),
            status: order.status,
            total: order.total,
        };
    }

    private async _validateProducts(input: PlaceOrderInputDto): Promise<void> {
        if (input.products.length === 0) {
            throw new Error("No products selected");
        }

        for (const product of input.products) {
            const p = await this._productFacade.checkStock({
                productId: product.productId,
            });
            if (p.stock <= 0) {
                throw new Error(`Product ${product.productId} is not available in stock`);
            }
        }
    }

    private async _getProduct(productId: string): Promise<Product> {
        const product = await this._catalogFacade.find({id: productId});
        if (!product) {
            throw new Error("Product not found");
        }
        const productProps = {
            id: new Id(product.id),
            name: product.name,
            description: product.description,
            salePrice: product.salePrice,
        };

        return new Product(productProps);
    }

}